# gen/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
...
  /user-page/advertisements/{id}:
    delete:
      operationId: deleteUserAdvertisement
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - user-page-controller
      x-accepts: application/json
    get:
      operationId: getUserAdvertisementById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdvertisementDto'
          description: OK
      tags:
      - user-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateUserAdvertisement
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdvertisementDto'
          description: OK
      tags:
      - user-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /advertisements/{id}:
    delete:
      operationId: deleteAdvertisement
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - advertisement-controller
      x-accepts: application/json
    get:
      operationId: getAdvertisementById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Advertisement'
          description: OK
      tags:
      - advertisement-controller
      x-accepts: '*/*'
    put:
      operationId: updateAdvertisement
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Advertisement'
          description: OK
      tags:
      - advertisement-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /admin-page/advertisements/{id}:
    delete:
      operationId: deleteAdvertisement_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - admin-page-controller
      x-accepts: application/json
    get:
      operationId: getAdvertisementById_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdvertisementDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateAdvertisement_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdvertisementDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /api/email/reset-password:
    post:
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - email-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /api/email/reset-password-request:
    post:
      operationId: resetPasswordRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - email-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /advertisements:
    get:
      operationId: getAdvertisements
      parameters:
      - explode: true
        in: query
        name: page
        required: true
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 50
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageAdvertisementDto'
          description: OK
      tags:
      - advertisement-controller
      x-accepts: '*/*'
    post:
      operationId: createAdvertisement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Advertisement'
          description: OK
      tags:
      - advertisement-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /admin-page/advertisements:
    get:
      operationId: getAllAdvertisements
      parameters:
      - explode: true
        in: query
        name: pageable
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageAdvertisementDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    post:
      operationId: createAdvertisement_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdvertisementDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /api/email/verify:
    get:
      operationId: verifyEmail
      parameters:
      - explode: true
        in: query
        name: token
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
          description: OK
      tags:
      - email-controller
      x-accepts: '*/*'
...
components:
...
  schemas:
...
    AdvertisementDto:
      example:
        id: 0
        title: title
        userId: 6
      properties:
        id:
          format: int64
          type: integer
        title:
          maxLength: 100
          minLength: 10
          type: string
        userId:
          format: int64
          type: integer
      required:
      - title
      type: object
    Advertisement:
      example:
        id: 0
        title: title
        user:
          favorites:
          - 1
          - 1
          advertisements:
          - null
          - null
          password: password
          role: ADMIN
          credentialsNonExpired: true
          accountNonExpired: true
          id: 6
          email: email
          enabled: true
          authorities:
          - authority: authority
          - authority: authority
          username: username
          accountNonLocked: true
      properties:
        id:
          format: int64
          type: integer
        title:
          type: string
        user:
          $ref: '#/components/schemas/User'
      type: object
    GrantedAuthority:
      example:
        authority: authority
      properties:
        authority:
          type: string
      type: object
    User:
      example:
        favorites:
        - 1
        - 1
        advertisements:
        - null
        - null
        password: password
        role: ADMIN
        credentialsNonExpired: true
        accountNonExpired: true
        id: 6
        email: email
        enabled: true
        authorities:
        - authority: authority
        - authority: authority
        username: username
        accountNonLocked: true
      properties:
        id:
          format: int64
          type: integer
        advertisements:
          items:
            $ref: '#/components/schemas/Advertisement'
          type: array
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          enum:
          - ADMIN
          - USER
          - GUEST
          type: string
        enabled:
          type: boolean
        favorites:
          items:
            format: int64
            type: integer
          type: array
        authorities:
          items:
            $ref: '#/components/schemas/GrantedAuthority'
          type: array
        accountNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
        credentialsNonExpired:
          type: boolean
      type: object
...
    EmailDto:
      example:
        newPassword: newPassword
        email: email
        token: token
      properties:
        token:
          type: string
        newPassword:
          type: string
        email:
          type: string
      type: object
...
    PageAdvertisementDto:
      example:
        number: 5
        size: 1
        last: true
        numberOfElements: 5
        totalPages: 0
        pageable:
          paged: true
          pageNumber: 7
          offset: 2
          pageSize: 9
          unpaged: true
          sort:
          - nullHandling: nullHandling
            ignoreCase: true
            property: property
            ascending: true
            direction: direction
          - nullHandling: nullHandling
            ignoreCase: true
            property: property
            ascending: true
            direction: direction
        sort:
        - nullHandling: nullHandling
          ignoreCase: true
          property: property
          ascending: true
          direction: direction
        - nullHandling: nullHandling
          ignoreCase: true
          property: property
          ascending: true
          direction: direction
        content:
        - id: 0
          title: title
          userId: 6
        - id: 0
          title: title
          userId: 6
        first: true
        totalElements: 6
        empty: true
      properties:
        totalPages:
          format: int32
          type: integer
        totalElements:
          format: int64
          type: integer
        size:
          format: int32
          type: integer
        content:
          items:
            $ref: '#/components/schemas/AdvertisementDto'
          type: array
        number:
          format: int32
          type: integer
        sort:
          items:
            $ref: '#/components/schemas/SortObject'
          type: array
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          format: int32
          type: integer
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
      type: object
    PageableObject:
      example:
        paged: true
        pageNumber: 7
        offset: 2
        pageSize: 9
        unpaged: true
        sort:
        - nullHandling: nullHandling
          ignoreCase: true
          property: property
          ascending: true
          direction: direction
        - nullHandling: nullHandling
          ignoreCase: true
          property: property
          ascending: true
          direction: direction
      properties:
        unpaged:
          type: boolean
        offset:
          format: int64
          type: integer
        sort:
          items:
            $ref: '#/components/schemas/SortObject'
          type: array
        pageNumber:
          format: int32
          type: integer
        pageSize:
          format: int32
          type: integer
        paged:
          type: boolean
      type: object
    SortObject:
      example:
        nullHandling: nullHandling
        ignoreCase: true
        property: property
        ascending: true
        direction: direction
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
      type: object
...
    Pageable:
      properties:
        page:
          format: int32
          minimum: 0
          type: integer
        size:
          format: int32
          minimum: 1
          type: integer
        sort:
          items:
            type: string
          type: array
      type: object
...